Delete my-app
