--- src/style.css
@@ -1,3 +1,0 @@
-body {
-  color: blue;
-}
